requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
bulma==1.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==39.0.2
cssselect2==0.7.0
distlib==0.3.6
Django==4.1.7
django-bulma==0.8.3
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-js-asset==2.0.0
filelock==3.9.0
html2pdf==0.1
html5lib==1.1
idna==3.4
lxml==4.9.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
oscrypto==1.3.0
pdfkit==1.0.0
Pillow==9.4.0
platformdirs==3.1.1
pycparser==2.21
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pypdf==3.5.1
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.14
virtualenv==20.20.0
webencodings==0.5.1
xhtml2pdf==0.2.9
